backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# File processing
PyPDF2>=3.0.0
pymupdf>=1.24.0
pymupdf4llm>=0.0.10
python-multipart>=0.0.6

# Google Drive integration
google-api-python-client>=2.100.0
google-auth>=2.20.0

# SSE support
sse-starlette>=1.6.0

# Database support (optional - for PostgreSQL/MySQL instead of JSON)
sqlalchemy>=2.0.0
pymysql>=1.1.0

# LangChain tools (lightweight - without heavy ML dependencies)
langchain>=0.1.0
langchain-community>=0.0.20
wikipedia>=1.4.0
arxiv>=2.0.0
yfinance>=0.2.30
duckduckgo-search>=4.0.0
tavily-python>=0.3.0
exa-py>=1.0.0

# TOON token optimization
python-toon>=0.1.0
tiktoken>=0.5.0

# Memory dependencies (optional - for embeddings-based memory)
# Uncomment these if you want memory feature and have enough disk space:
# sentence-transformers>=2.2.0
# chromadb>=0.4.0
